fix: 각 구분자 별 소수가 포함된 경우 예외 처리

+++ src/main/java/calculator/Validator.java
@@ -6,7 +6,7 @@ public class Validator {
     public static void validate(List<String> numbers) {
         for (String number : numbers) {
             try {
-                int n = Integer.parseInt(number);
+                float n = Float.parseFloat(number);
                 if (n < 0) throw new InvalidNumberFormatException();
             } catch (NumberFormatException e) {
                 throw new InvalidNumberFormatException();
